requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
albumentations==1.4.0
    # via insightface
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
coloredlogs==15.0.1
    # via onnxruntime-gpu
contourpy==1.2.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
cython==3.0.8
    # via insightface
easydict==1.12
    # via insightface
filelock==3.13.1
    # via
    #   huggingface-hub
    #   torch
    #   triton
flatbuffers==23.5.26
    # via onnxruntime-gpu
fonttools==4.49.0
    # via matplotlib
fsspec==2024.2.0
    # via
    #   huggingface-hub
    #   torch
hsemotion==0.3.0
    # via -r requirements.in
huggingface-hub==0.21.3
    # via timm
humanfriendly==10.0
    # via coloredlogs
idna==3.6
    # via requests
imageio==2.34.0
    # via scikit-image
insightface==0.7.3
    # via -r requirements.in
jinja2==3.1.3
    # via torch
joblib==1.3.2
    # via scikit-learn
kiwisolver==1.4.5
    # via matplotlib
lazy-loader==0.3
    # via scikit-image
markupsafe==2.1.5
    # via jinja2
matplotlib==3.8.3
    # via insightface
mpmath==1.3.0
    # via sympy
networkx==3.2.1
    # via
    #   scikit-image
    #   torch
numpy==1.26.4
    # via
    #   -r requirements.in
    #   albumentations
    #   contourpy
    #   hsemotion
    #   imageio
    #   insightface
    #   matplotlib
    #   onnx
    #   onnxruntime-gpu
    #   opencv-python
    #   opencv-python-headless
    #   pandas
    #   qudida
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   tifffile
    #   torchvision
nvidia-cublas-cu12==12.1.3.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.1.105
    # via torch
nvidia-cuda-nvrtc-cu12==12.1.105
    # via torch
nvidia-cuda-runtime-cu12==12.1.105
    # via torch
nvidia-cudnn-cu12==8.9.2.26
    # via torch
nvidia-cufft-cu12==11.0.2.54
    # via torch
nvidia-curand-cu12==10.3.2.106
    # via torch
nvidia-cusolver-cu12==11.4.5.107
    # via torch
nvidia-cusparse-cu12==12.1.0.106
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.18.1
    # via torch
nvidia-nvjitlink-cu12==12.3.101
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
nvidia-nvtx-cu12==12.1.105
    # via torch
onnx==1.15.0
    # via insightface
onnxruntime-gpu==1.17.1
    # via -r requirements.in
opencv-python==4.9.0.80
    # via
    #   -r requirements.in
    #   albumentations
opencv-python-headless==4.9.0.80
    # via
    #   -r requirements.in
    #   qudida
packaging==23.2
    # via
    #   huggingface-hub
    #   matplotlib
    #   onnxruntime-gpu
    #   scikit-image
pandas==2.2.1
    # via -r requirements.in
pillow==10.2.0
    # via
    #   hsemotion
    #   imageio
    #   insightface
    #   matplotlib
    #   scikit-image
    #   torchvision
prettytable==3.10.0
    # via insightface
protobuf==4.25.3
    # via
    #   onnx
    #   onnxruntime-gpu
pyparsing==3.1.1
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   albumentations
    #   huggingface-hub
    #   timm
qudida==0.0.4
    # via albumentations
requests==2.31.0
    # via
    #   huggingface-hub
    #   insightface
    #   torchvision
safetensors==0.4.2
    # via timm
scikit-image==0.22.0
    # via
    #   -r requirements.in
    #   albumentations
    #   insightface
scikit-learn==1.4.1.post1
    # via
    #   insightface
    #   qudida
scipy==1.12.0
    # via
    #   albumentations
    #   insightface
    #   scikit-image
    #   scikit-learn
six==1.16.0
    # via python-dateutil
sympy==1.12
    # via
    #   onnxruntime-gpu
    #   torch
threadpoolctl==3.3.0
    # via scikit-learn
tifffile==2024.2.12
    # via scikit-image
timm==0.9.16
    # via hsemotion
torch==2.1.2
    # via
    #   -r requirements.in
    #   hsemotion
    #   timm
    #   torchvision
torchvision==0.16.2
    # via
    #   -r requirements.in
    #   hsemotion
    #   timm
tqdm==4.65.0
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   insightface
triton==2.1.0
    # via torch
typing-extensions==4.10.0
    # via
    #   huggingface-hub
    #   qudida
    #   torch
tzdata==2024.1
    # via pandas
urllib3==2.2.1
    # via requests
wcwidth==0.2.13
    # via prettytable
